# docs/swagger.yaml
definitions:
...
  handler.InvoicePage:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Invoice'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        format: int64
        type: integer
    type: object
...
  model.Attachment:
    properties:
      createdAt:
        type: string
      id:
        type: string
      mimeType:
        type: string
      model:
        type: string
      modelID:
        type: string
      originalName:
        type: string
      serverName:
        type: string
      size:
        type: integer
      updatedAt:
        type: string
    type: object
  model.Client:
    properties:
      address:
        type: string
      clientType:
        type: string
      contractType:
        type: string
      createdAt:
        type: string
      description:
        type: string
      finishedAt:
        type: string
      hourRate:
        type: number
      id:
        type: string
      name:
        type: string
      startedAt:
        type: string
      tags:
        items:
          type: string
        type: array
      taxNumber:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Invoice:
    properties:
      billableHours:
        type: number
      client:
        $ref: '#/definitions/model.Client'
      createdAt:
        type: string
      deliveredAt:
        type: string
      dueAt:
        type: string
      externalId:
        type: string
      externalLink:
        type: string
      id:
        type: string
      isCanceled:
        type: boolean
      issuedAt:
        type: string
      name:
        type: string
      note:
        type: string
      paidAt:
        type: string
      total:
        type: number
      updatedAt:
        type: string
    type: object
  model.InvoiceRequest:
    properties:
      billableHours:
        type: number
      clientId:
        type: string
      deliveredAt:
        type: string
      dueAt:
        type: string
      externalId:
        type: string
      externalLink:
        type: string
      isCanceled:
        type: boolean
      issuedAt:
        type: string
      name:
        type: string
      note:
        type: string
      paidAt:
        type: string
      total:
        type: number
    type: object
  model.KanbanBoard:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastActivityAt:
        type: string
      name:
        type: string
      projectId:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/inri/invoices:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of all invoices
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: pageSize
        type: integer
      - description: RSQL filter query
        in: query
        name: filter
        type: string
      - description: Sort query
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.InvoicePage'
      security:
      - GatewayAuth: []
      tags:
      - Invoices
    post:
      consumes:
      - application/json
      description: Creates a new invoice
      parameters:
      - description: Invoice data
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Invoice'
      security:
      - GatewayAuth: []
      tags:
      - Invoices
  /v1/inri/invoices/{invoiceId}:
    delete:
      consumes:
      - application/json
      description: Deletes the invoice
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - Invoices
    get:
      consumes:
      - application/json
      description: Returns an invoice by its ID
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Invoice'
      security:
      - GatewayAuth: []
      tags:
      - Invoices
    put:
      consumes:
      - application/json
      description: Updates an existing invoice
      parameters:
      - description: Invoice data
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceRequest'
      - description: Invoice ID
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Invoice'
      security:
      - GatewayAuth: []
      tags:
      - Invoices
  /v1/inri/invoices/{invoiceId}/attachments:
    get:
      consumes:
      - application/json
      description: Returns a list of attachments by invoice ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Attachment'
            type: array
      security:
      - GatewayAuth: []
      tags:
      - Invoices
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple files as attachments for the given invoice
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceId
        required: true
        type: string
      - collectionFormat: csv
        description: Files to upload
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - Invoices
...
  /v1/inri/projects/{projectId}/kanban:
    get:
      consumes:
      - application/json
      description: Returns all boards of a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.KanbanBoard'
            type: array
      security:
      - GatewayAuth: []
      tags:
      - Kanban boards
    post:
      consumes:
      - application/json
      description: Creates a new Kanban board for a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.KanbanBoard'
      security:
      - GatewayAuth: []
      tags:
      - Kanban boards
  /v1/inri/projects/{projectId}/kanban/{kanbanId}:
    delete:
      consumes:
      - application/json
      description: Deletes a Kanban board
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - Kanban boards
    get:
      consumes:
      - application/json
      description: Returns a single Kanban board
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KanbanBoard'
      security:
      - GatewayAuth: []
      tags:
      - Kanban boards
    put:
      consumes:
      - application/json
      description: Updates the Kanban board
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Kanban ID
        in: path
        name: kanbanId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.KanbanBoard'
      security:
      - GatewayAuth: []
      tags:
      - Kanban boards
...
securityDefinitions:
  GatewayAuth:
    description: Injected by API Gateway. Do not use in production.
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"